refactor(metadata): migrate host component to v2
