+++ cmd/agent/subcommands/flare/command.go
@@ -58,7 +58,7 @@ import (
 	haagentfx "github.com/DataDog/datadog-agent/comp/haagent/fx"
 	healthplatform "github.com/DataDog/datadog-agent/comp/healthplatform"
 	haagentmetadatafx "github.com/DataDog/datadog-agent/comp/metadata/haagent/fx"
-	"github.com/DataDog/datadog-agent/comp/metadata/host/hostimpl"
+	hostfx "github.com/DataDog/datadog-agent/comp/metadata/host/fx"
 	inventoryagentfx "github.com/DataDog/datadog-agent/comp/metadata/inventoryagent/fx"
 	"github.com/DataDog/datadog-agent/comp/metadata/inventoryhost/inventoryhostimpl"
 	resourcesfx "github.com/DataDog/datadog-agent/comp/metadata/resources/fx"
@@ -159,7 +159,7 @@ func Commands(globalParams *command.GlobalParams) []*cobra.Command {
 				healthplatform.Bundle(),
 				// We need inventoryagent to fill the status page generated by the flare.
 				inventoryagentfx.Module(),
-				hostimpl.Module(),
+				hostfx.Module(),
 				inventoryhostimpl.Module(),
 				haagentmetadatafx.Module(),
 				resourcesfx.Module(),
