+++ comp/metadata/host/def/component.go
