+++ comp/metadata/host/impl/hosttags/tags_test.go
