+++ comp/metadata/host/impl/payload.go
@@ -11,7 +11,7 @@ import (
 	"context"
 	"encoding/json"
 
-	"github.com/DataDog/datadog-agent/comp/metadata/host/hostimpl/utils"
+	"github.com/DataDog/datadog-agent/comp/metadata/host/impl/utils"
 	"github.com/DataDog/datadog-agent/pkg/config/env"
 	"github.com/DataDog/datadog-agent/pkg/gohai"
 )
