+++ comp/metadata/host/impl/status_templates/host.tmpl
