+++ comp/metadata/host/impl/utils/common.go
