+++ comp/metadata/host/impl/utils/common_test.go
