+++ comp/metadata/host/impl/utils/host_test.go
