+++ comp/metadata/host/impl/utils/host_windows.go
