+++ comp/metadata/host/impl/utils/host_windows_test.go
