+++ comp/metadata/host/impl/utils/meta.go
